Quick benchmark:
```
$ git checkout main
Switched to branch 'main'
$ JAX_NUM_GENERATED_CASES=20 pytest -n auto tests/sparse_test.py -k broadcast_in_dim --durations=20
=============================================== test session starts ================================================
platform darwin -- Python 3.8.2, pytest-7.1.2, pluggy-0.13.1
rootdir: /Users/vanderplas/github/google/jax, configfile: pytest.ini
plugins: xdist-2.5.0, env-0.6.2, cov-2.10.0, forked-1.3.0, reportlog-0.1.2
gw0 [20] / gw1 [20] / gw2 [20] / gw3 [20] / gw4 [20] / gw5 [20] / gw6 [20] / gw7 [20]
....................                                                                                         [100%]
=============================================== slowest 20 durations ===============================================
5.97s call     tests/sparse_test.py::BCOOTest::test_bcoo_broadcast_in_dim15
3.78s call     tests/sparse_test.py::BCOOTest::test_bcoo_broadcast_in_dim11
3.72s call     tests/sparse_test.py::BCOOTest::test_bcoo_broadcast_in_dim12
3.09s call     tests/sparse_test.py::BCOOTest::test_bcoo_broadcast_in_dim14
3.03s call     tests/sparse_test.py::BCOOTest::test_bcoo_broadcast_in_dim9
3.02s call     tests/sparse_test.py::BCOOTest::test_bcoo_broadcast_in_dim1
2.39s call     tests/sparse_test.py::BCOOTest::test_bcoo_broadcast_in_dim3
2.22s call     tests/sparse_test.py::BCOOTest::test_bcoo_broadcast_in_dim16
2.22s call     tests/sparse_test.py::BCOOTest::test_bcoo_broadcast_in_dim13
2.18s call     tests/sparse_test.py::BCOOTest::test_bcoo_broadcast_in_dim5
1.91s call     tests/sparse_test.py::BCOOTest::test_bcoo_broadcast_in_dim7
1.88s call     tests/sparse_test.py::BCOOTest::test_bcoo_broadcast_in_dim19
1.75s call     tests/sparse_test.py::BCOOTest::test_bcoo_broadcast_in_dim10
1.70s call     tests/sparse_test.py::BCOOTest::test_bcoo_broadcast_in_dim8
1.55s call     tests/sparse_test.py::BCOOTest::test_bcoo_broadcast_in_dim2
1.17s call     tests/sparse_test.py::BCOOTest::test_bcoo_broadcast_in_dim4
1.01s call     tests/sparse_test.py::BCOOTest::test_bcoo_broadcast_in_dim6
0.96s call     tests/sparse_test.py::BCOOTest::test_bcoo_broadcast_in_dim0
0.90s call     tests/sparse_test.py::BCOOTest::test_bcoo_broadcast_in_dim17
0.29s call     tests/sparse_test.py::BCOOTest::test_bcoo_broadcast_in_dim18
=============================================== 20 passed in 12.12s ================================================
$ git checkout bcoo-broadcast-performance
Switched to branch 'bcoo-broadcast-performance'
$ JAX_NUM_GENERATED_CASES=20 pytest -n auto tests/sparse_test.py -k broadcast_in_dim --durations=20
=============================================== test session starts ================================================
platform darwin -- Python 3.8.2, pytest-7.1.2, pluggy-0.13.1
rootdir: /Users/vanderplas/github/google/jax, configfile: pytest.ini
plugins: xdist-2.5.0, env-0.6.2, cov-2.10.0, forked-1.3.0, reportlog-0.1.2
gw0 [20] / gw1 [20] / gw2 [20] / gw3 [20] / gw4 [20] / gw5 [20] / gw6 [20] / gw7 [20]
....................                                                                                         [100%]
=============================================== slowest 20 durations ===============================================
4.80s call     tests/sparse_test.py::BCOOTest::test_bcoo_broadcast_in_dim15
2.93s call     tests/sparse_test.py::BCOOTest::test_bcoo_broadcast_in_dim11
2.85s call     tests/sparse_test.py::BCOOTest::test_bcoo_broadcast_in_dim9
2.84s call     tests/sparse_test.py::BCOOTest::test_bcoo_broadcast_in_dim12
2.20s call     tests/sparse_test.py::BCOOTest::test_bcoo_broadcast_in_dim14
2.13s call     tests/sparse_test.py::BCOOTest::test_bcoo_broadcast_in_dim3
2.12s call     tests/sparse_test.py::BCOOTest::test_bcoo_broadcast_in_dim1
2.04s call     tests/sparse_test.py::BCOOTest::test_bcoo_broadcast_in_dim5
1.84s call     tests/sparse_test.py::BCOOTest::test_bcoo_broadcast_in_dim16
1.71s call     tests/sparse_test.py::BCOOTest::test_bcoo_broadcast_in_dim19
1.70s call     tests/sparse_test.py::BCOOTest::test_bcoo_broadcast_in_dim7
1.63s call     tests/sparse_test.py::BCOOTest::test_bcoo_broadcast_in_dim8
1.40s call     tests/sparse_test.py::BCOOTest::test_bcoo_broadcast_in_dim2
1.39s call     tests/sparse_test.py::BCOOTest::test_bcoo_broadcast_in_dim13
0.98s call     tests/sparse_test.py::BCOOTest::test_bcoo_broadcast_in_dim4
0.97s call     tests/sparse_test.py::BCOOTest::test_bcoo_broadcast_in_dim10
0.93s call     tests/sparse_test.py::BCOOTest::test_bcoo_broadcast_in_dim6
0.67s call     tests/sparse_test.py::BCOOTest::test_bcoo_broadcast_in_dim17
0.36s call     tests/sparse_test.py::BCOOTest::test_bcoo_broadcast_in_dim0
0.25s call     tests/sparse_test.py::BCOOTest::test_bcoo_broadcast_in_dim18
=============================================== 20 passed in 11.26s ================================================